+++ src/mesa/main/texstore.c
@@ -2127,7 +2127,7 @@ _mesa_validate_pbo_compressed_teximage(GLcontext *ctx,
       return pixels;
    }
    if ((const GLubyte *) pixels + imageSize >
-       (const GLubyte *) packing->BufferObj->Size) {
+       (const GLubyte *)(uintptr_t) packing->BufferObj->Size) {
       /* out of bounds read! */
       _mesa_error(ctx, GL_INVALID_OPERATION, funcName, "(invalid PBO access");
       return NULL;
