+++ src/mesa/shader/grammar_mesa.c
@@ -40,17 +40,17 @@ void grammar_alloc_free (void *ptr)
     _mesa_free (ptr);
 }
 
-void *grammar_alloc_malloc (unsigned int size)
+void *grammar_alloc_malloc (size_t size)
 {
     return _mesa_malloc (size);
 }
 
-void *grammar_alloc_realloc (void *ptr, unsigned int old_size, unsigned int size)
+void *grammar_alloc_realloc (void *ptr, size_t old_size, size_t size)
 {
     return _mesa_realloc (ptr, old_size, size);
 }
 
-void *grammar_memory_copy (void *dst, const void * src, unsigned int size)
+void *grammar_memory_copy (void *dst, const void * src, size_t size)
 {
     return _mesa_memcpy (dst, src, size);
 }
@@ -60,7 +60,7 @@ int grammar_string_compare (const byte *str1, const byte *str2)
     return _mesa_strcmp ((const char *) str1, (const char *) str2);
 }
 
-int grammar_string_compare_n (const byte *str1, const byte *str2, unsigned int n)
+int grammar_string_compare_n (const byte *str1, const byte *str2, size_t n)
 {
     return _mesa_strncmp ((const char *) str1, (const char *) str2, n);
 }
@@ -70,7 +70,7 @@ byte *grammar_string_copy (byte *dst, const byte *src)
     return (byte *) _mesa_strcpy ((char *) dst, (const char *) src);
 }
 
-byte *grammar_string_copy_n (byte *dst, const byte *src, unsigned int n)
+byte *grammar_string_copy_n (byte *dst, const byte *src, size_t n)
 {
     return (byte *) _mesa_strncpy ((char *) dst, (const char *) src, n);
 }
@@ -82,6 +82,6 @@ byte *grammar_string_duplicate (const byte *src)
 
 unsigned int grammar_string_length (const byte *str)
 {
-    return _mesa_strlen ((const char *) str);
+    return (unsigned int)_mesa_strlen ((const char *) str);
 }
